# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API StageSeller",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Api StageSeller"
        }
    ],
    "paths": {
...
        "/api/stages": {
            "get": {
                "tags": [
                    "Stage"
                ],
                "operationId": "1681ed7b07c2407752bcebd23c353695",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted / add stage",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Stage-add"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid JWT Token code || Expired JWT Token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/stages/{id}": {
            "put": {
                "tags": [
                    "Stage"
                ],
                "operationId": "697e496615112ea9b90c6a2a1d6a139b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted / edit stage",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Stage-edit"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid JWT Token code || Expired JWT Token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Stage-add": {
                "description": "Stage-add",
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "duration": {
                        "type": "integer"
                    },
                    "location": {
                        "type": "string"
                    },
                    "isRemoteFriendly": {
                        "type": "boolean"
                    },
                    "isTravelFriendly": {
                        "type": "boolean"
                    },
                    "skills": {
                        "type": "array"
                    }
                },
                "type": "object"
            },
            "Stage-edit": {
                "description": "Stage-edit",
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "duration": {
                        "type": "integer"
                    },
                    "location": {
                        "type": "string"
                    },
                    "isRemoteFriendly": {
                        "type": "boolean"
                    },
                    "isTravelFriendly": {
                        "type": "boolean"
                    },
                    "skills": {
                        "type": "array"
                    }
                },
                "type": "object"
            },
...
        }
    }
}